Move the MovesProvider into App.js
